dev/requirements.txt
weibo
google-cloud-language
request
webbrowser